ABM tables finished

+++ src/app/shared/components/table-component/table-component.ts
@@ -53,17 +53,20 @@ export class TableComponent {
   }
   deleteTable(table: Table){
     console.log("mesa n: " + table.number);
-    if(table.order || table.status === 'OCCUPIED'){
-      alert("La mesa no se puede eliminar porque tiene una orden asignada.");
-      return;
-    }
-
-    this._tableStore.deleteTableInDb(table.number).subscribe(
-      {
-        next: () => {alert("Mesa eliminada satisfactoriamente")},
-        error: (err) => {alert(err.error.message)}
+    const res = confirm("¿Desea eliminar la mesa?");
+    if(res){
+      if(table.order || table.status === 'OCCUPIED'){
+        alert("La mesa no se puede eliminar porque tiene una orden asignada.");
+        return;
       }
-    );
+  
+      this._tableStore.deleteTableInDb(table.number).subscribe(
+        {
+          next: () => {alert("Mesa eliminada satisfactoriamente")},
+          error: (err) => {alert(err.error.message)}
+        }
+      );
+    }
   }
 
   reserveTable(): void {
